Add comments to lib/web.js
Closes issue #7.

- Add comments to all modules
- Move util helper methods to the modules that use them
