+++ lib/auth.js
@@ -29,7 +29,10 @@ If you would like to know more about the solid Solid project, please see
 https://github.com/solid/
 */
 
-// WebID authentication and signup
+/**
+ * Provides Solid methods for WebID authentication and signup
+ * @module auth
+ */
 'use strict'
 var XMLHttpRequest = require('xhr2')
 
@@ -37,7 +40,12 @@ var XMLHttpRequest = require('xhr2')
 var authEndpoint = 'https://databox.me/'
 var signupEndpoint = 'https://solid.github.io/solid-idps/'
 
-// Listen to login messages from child window/iframe
+/**
+ * Sets up an event listener to monitor login messages from child window/iframe
+ * @method listen
+ * @static
+ * @return {Promise<String>} Event listener promise, resolves to user's WebID
+ */
 var listen = function listen () {
   var promise = new Promise(function (resolve, reject) {
     var eventMethod = window.addEventListener
@@ -63,9 +71,16 @@ var listen = function listen () {
   return promise
 }
 
-// attempt to find the current user's WebID from the User header
-// if authenticated
-// resolve(webid) - string
+/**
+ * Performs a Solid login() via an XHR HEAD operation.
+ * (Attempts to find the current user's WebID from the User header, if
+ *   already authenticated.)
+ * @method login
+ * @static
+ * @param url {String} Location of a Solid server or container at which the
+ *   user might be authenticated.
+ * @return {Promise<String>} XHR HEAD op promise, resolves to user's WebID
+ */
 var login = function login (url) {
   url = url || window.location.origin + window.location.pathname
   var promise = new Promise(function (resolve, reject) {
@@ -104,7 +119,14 @@ var login = function login (url) {
   return promise
 }
 
-// Open signup window
+/**
+ * Opens a signup popup window, sets up `listen()`.
+ * @method signup
+ * @static
+ * @param url {String} Location of a Solid server for user signup.
+ * @return {Promise<String>} Returns a listener promise, resolves with signed
+ *   up user's WebID.
+ */
 var signup = function signup (url) {
   url = url || signupEndpoint
   var leftPosition, topPosition
