+++ lib/identity.js
@@ -1,7 +1,10 @@
 'use strict'
-// Identity / WebID
+/**
+ * Provides Solid helper functions involved with parsing a user's WebId profile.
+ * @module identity
+ */
+
 var solidClient = require('./web')
-var appendGraph = require('./utils').appendGraph
 
 // common vocabs
 // var RDF = $rdf.Namespace('http://www.w3.org/1999/02/22-rdf-syntax-ns#')
@@ -10,8 +13,29 @@ var PIM = $rdf.Namespace('http://www.w3.org/ns/pim/space#')
 var FOAF = $rdf.Namespace('http://xmlns.com/foaf/0.1/')
 var DCT = $rdf.Namespace('http://purl.org/dc/terms/')
 
-// fetch user profile (follow sameAs links) and return promise with a graph
-// resolve(graph)
+/**
+ * Appends RDF statements from one graph object to another
+ * @method appendGraph
+ * @param toGraph {Graph} $rdf.Graph object to append to
+ * @param fromGraph {Graph} $rdf.Graph object to append from
+ * @param docURI {String} Document URI to use as source
+ */
+var appendGraph = function appendGraph (toGraph, fromGraph, docURI) {
+  var source = (docURI) ? $rdf.sym(docURI) : undefined
+  fromGraph.statementsMatching(undefined, undefined, undefined, source)
+    .forEach(function (st) {
+      toGraph.add(st.subject, st.predicate, st.object, st.why)
+    })
+}
+
+/**
+ * Fetches a user's WebId profile, follows `sameAs` links,
+ *   and return a promise with a parsed RDF graph of the results.
+ * @method getProfile
+ * @static
+ * @param url {String} WebId or Location of a user's profile.
+ * @return {Promise<Graph>}
+ */
 var getProfile = function getProfile (url) {
   var promise = new Promise(function (resolve, reject) {
     // Load main profile
@@ -66,8 +90,15 @@ var getProfile = function getProfile (url) {
   return promise
 }
 
-// Find the user's workspaces
-// Return an object with the list of objects (workspaces)
+/**
+ * Finds the Workspaces linked from the user's WebId Profile.
+ * (Optionally fetches the profile, if it hasn't already been loaded.)
+ * @method getWorkspaces
+ * @static
+ * @param webid {String} WebId or Location of a user's profile.
+ * @param graph {Graph} Parsed graph of the user's profile.
+ * @return {Array<Object>} List of parsed Workspace triples.
+ */
 var getWorkspaces = function getWorkspaces (webid, graph) {
   var promise = new Promise(function (resolve, reject) {
     if (!graph) {
@@ -108,8 +139,14 @@ var getWorkspaces = function getWorkspaces (webid, graph) {
   return promise
 }
 
-// Find the user's writable profiles
-// Return an object with the list of profile URIs
+/**
+ * Finds writeable profiles linked from the user's WebId Profile.
+ * @method getWritableProfiles
+ * @static
+ * @param webid {String} WebId or Location of a user's profile.
+ * @param graph {Graph} Parsed graph of the user's profile.
+ * @return {Array<Object>} List of writeable profile triples
+ */
 var getWritableProfiles = function getWritableProfiles (webid, graph) {
   var promise = new Promise(function (resolve, reject) {
     if (!graph) {
