+++ lib/status.js
@@ -1,21 +1,39 @@
 'use strict'
+/**
+ * Provides Web API helpers dealing with a user's online / offline status.
+ * @module status
+ */
 
-// Get current online status
+/**
+ * Returns a user's online status (true if user is on line)
+ * @method isOnline
+ * @static
+ * @return {Boolean}
+ */
 var isOnline = function isOnline () {
   return window.navigator.onLine
 }
 
-// Is offline
+/**
+ * Adds an even listener to trigger when the user goes offline.
+ * @method onOffline
+ * @static
+ * @param callback {Function} Callback to invoke when user goes offline.
+ */
 var onOffline = function onOffline (callback) {
   window.addEventListener('offline', callback, false)
 }
 
-// Trigger when user comes online
+/**
+ * Adds an even listener to trigger when the user comes online.
+ * @method onOnline
+ * @static
+ * @param callback {Function} Callback to invoke when user comes online
+ */
 var onOnline = function onOnline (callback) {
   window.addEventListener('online', callback, false)
 }
 
-// Events
 module.exports.isOnline = isOnline
 module.exports.onOffline = onOffline
 module.exports.onOnline = onOnline
