+++ lib/web.js
@@ -1,7 +1,10 @@
 'use strict'
-var XMLHttpRequest = require('xhr2')
-var parseLinkHeader = require('./utils').parseLinkHeader
+/**
+ * Provides a Solid web client class for performing LDP CRUD operations.
+ * @module web
+ */
 
+var XMLHttpRequest = require('xhr2')
 // Init some defaults
 var PROXY = 'https://databox.me/,proxy?uri={uri}'
 var TIMEOUT = 5000
@@ -10,7 +13,40 @@ $rdf.Fetcher.crossSiteProxyTemplate = PROXY
 // common vocabs
 var LDP = $rdf.Namespace('http://www.w3.org/ns/ldp#')
 
-// return metadata for a given request
+/**
+ * Parses a Link header from an XHR HTTP Request.
+ * @method parseLinkHeader
+ * @param link {String} Contents of the Link response header
+ * @return {Object}
+ */
+var parseLinkHeader = function parseLinkHeader (link) {
+  var linkexp = /<[^>]*>\s*(\s*;\s*[^\(\)<>@,;:"\/\[\]\?={} \t]+=(([^\(\)<>@,;:"\/\[\]\?={} \t]+)|("[^"]*")))*(,|$)/g
+  var paramexp = /[^\(\)<>@,;:"\/\[\]\?={} \t]+=(([^\(\)<>@,;:"\/\[\]\?={} \t]+)|("[^"]*"))/g
+
+  var matches = link.match(linkexp)
+  var rels = {}
+  for (var i = 0; i < matches.length; i++) {
+    var split = matches[i].split('>')
+    var href = split[0].substring(1)
+    var ps = split[1]
+    var s = ps.match(paramexp)
+    for (var j = 0; j < s.length; j++) {
+      var p = s[j]
+      var paramsplit = p.split('=')
+      // var name = paramsplit[0]
+      var rel = paramsplit[1].replace(/["']/g, '')
+      rels[rel] = href
+    }
+  }
+  return rels
+}
+
+/**
+ * Parses an XHR response and composes a meta object
+ * @method parseResponseMeta
+ * @param resp {XMLHttpRequest} Result of an XHR
+ * @return {Object} Parsed response object
+ */
 function parseResponseMeta (resp) {
   var h = parseLinkHeader(resp.getResponseHeader('Link'))
   var meta = {}
@@ -49,11 +85,44 @@ function parseResponseMeta (resp) {
   return meta
 }
 
-// LDP operations
-var client = {
-  // check if a resource exists and return useful Solid info
-  // (acl, meta, type, etc)
-  // resolve(metaObj)
+/**
+ * Resolves a successful XHR (and returns a parsed meta object of the response),
+ *   or rejects a failed request.
+ * @method readyStateChange
+ * @param resolve {Function} Resolve callback
+ * @param reject {Function} Reject callback
+ * @param parseResponse {Boolean} Whether to parse a successful response
+ * @return {Object|XMLHttpRequest} If request is successful, returns a parsed
+ *   or unparsed response (depending on if `parseResponse` is set to true)
+ */
+var readyStateChange = function readyStateChange (xhr, resolve, reject,
+    parseResponse) {
+  if (xhr.readyState === xhr.DONE) {
+    if (xhr.status === 200 || xhr.status === 201) {
+      if (parseResponse) {
+        return resolve(parseResponseMeta(xhr))
+      } else {
+        return xhr
+      }
+    } else {
+      reject({status: xhr.status, xhr: xhr})
+    }
+  }
+}
+
+/**
+ * Providers a collection of Solid/LDP web operations (CRUD)
+ * @class SolidWebClient
+ * @static
+ */
+var SolidWebClient = {
+  /**
+   * Checks to see if a Solid resource exists, and returns useful resource
+   *   metadata info.
+   * @method head
+   * @param url {String} URL of a resource or container
+   * @return {Promise} Result of an HTTP HEAD operation (returns a meta object)
+   */
   head: function head (url) {
     var promise = new Promise(function (resolve) {
       var http = new XMLHttpRequest()
@@ -70,8 +139,12 @@ var client = {
     return promise
   },
 
-  // fetch an RDF resource
-  // resolve(graph) | reject(this)
+  /**
+   * Retrieves a resource or container by making an HTTP GET call.
+   * @method get
+   * @param url {String} URL of the resource or container to fetch
+   * @return {Promise|Object} Result of the HTTP GET operation
+   */
   get: function get (url) {
     var promise = new Promise(function (resolve, reject) {
       var g = $rdf.graph()
@@ -92,15 +165,26 @@ var client = {
     return promise
   },
 
-  // create new resource
-  // resolve(metaObj) | reject
+  /**
+   * Creates a new resource by performing
+   *   a Solid/LDP POST operation to a specified container.
+   * @param url {String} URL of the container to post to
+   * @param data {Object} Data/payload of the resource to be created
+   * @param slug {String} Suggested URL fragment for the new resource
+   * @param isContainer {Boolean} Is the object being created a Container
+   *            or Resource?
+   * @param mime {String} MIME Type
+   * @method post
+   * @return {Promise|Object} Result of XHR POST (returns parsed
+   *     response meta object) or an anonymous error object with status code
+   */
   post: function post (url, data, slug, isContainer, mime) {
     var resType = LDP('Resource').uri
     if (isContainer) {
       resType = LDP('BasicContainer').uri
       mime = 'text/turtle' // force right mime for containers only
     }
-    mime = (mime) ? mime : 'text/turtle'
+    mime = mime || 'text/turtle'
     var promise = new Promise(function (resolve, reject) {
       var http = new XMLHttpRequest()
       http.open('POST', url)
@@ -110,15 +194,7 @@ var client = {
         http.setRequestHeader('Slug', slug)
       }
       http.withCredentials = true
-      http.onreadystatechange = function () {
-        if (this.readyState === this.DONE) {
-          if (this.status === 200 || this.status === 201) {
-            resolve(parseResponseMeta(this))
-          } else {
-            reject({status: this.status, xhr: this})
-          }
-        }
-      }
+      http.onreadystatechange = readyStateChange(http, resolve, reject, true)
       if (data && data.length > 0) {
         http.send(data)
       } else {
@@ -129,11 +205,20 @@ var client = {
     return promise
   },
 
-  // update/create resource using HTTP PUT
-  // resolve(metaObj) | reject
+  /**
+   * Updates an existing resource or creates a new resource by performing
+   *   a Solid/LDP PUT operation to a specified container
+   * @method put
+   * @param url {String} URL of the resource to be updated/created
+   * @param data {Object} Data/payload of the resource to be created or updated
+   * @param mime {String} MIME Type of the resource to be created
+   * @return {Promise|Object} Result of PUT operation (returns parsed response
+   *     meta object if successful, rejects with an anonymous error status
+   *     object if not successful)
+   */
   put: function put (url, data, mime) {
     var promise = new Promise(function (resolve, reject) {
-      mime = (mime) ? mime : 'text/turtle'
+      mime = mime || 'text/turtle'
       var http = new XMLHttpRequest()
       http.open('PUT', url)
       http.setRequestHeader('Content-Type', mime)
@@ -147,6 +232,10 @@ var client = {
           }
         }
       }
+      // Handle network errors
+      http.onerror = function () {
+        reject(Error('Network Error'))
+      }
       if (data) {
         http.send(data)
       } else {
@@ -157,9 +246,19 @@ var client = {
     return promise
   },
 
-  // patch a resource
-  // accepts arrays of individual statements (turtle) as params
-  // e.g. [ '<a> <b> <c> .', '<d> <e> <f> .']
+  /**
+   * Partially edits an RDF-type resource by performing a PATCH operation.
+   *   Accepts arrays of individual statements (in Turtle format) as params.
+   *   For example:
+   *   [ '<a> <b> <c> .', '<d> <e> <f> .']
+   * @method patch
+   * @param url {String} URL of the resource to be edited
+   * @param toDel {Array<String>} Triples to remove from the resource
+   * @param toIns {Array<String>} Triples to insert into the resource
+   * @return {Promise|Object} Result of PATCH operation (returns parsed response
+   *     meta object if successful, rejects with an anonymous error status
+   *     object if not successful)
+   */
   patch: function patch (url, toDel, toIns) {
     var promise = new Promise(function (resolve, reject) {
       var data = ''
@@ -188,7 +287,7 @@ var client = {
       http.withCredentials = true
       http.onreadystatechange = function () {
         if (this.readyState === this.DONE) {
-          if (this.status === 200) {
+          if (this.status === 200 || this.status === 201) {
             return resolve(parseResponseMeta(this))
           } else {
             reject({status: this.status, xhr: this})
@@ -205,8 +304,13 @@ var client = {
     return promise
   },
 
-  // delete a resource
-  // resolve(true) | reject
+  /**
+   * Deletes an existing resource or container.
+   * @method del
+   * @param url {String} URL of the resource or container to be deleted
+   * @return {Promise|Object} Result of the HTTP Delete operation (returns true
+   *   on success, or an anonymous error object on failure)
+   */
   del: function del (url) {
     var promise = new Promise(function (resolve, reject) {
       var http = new XMLHttpRequest()
@@ -214,7 +318,7 @@ var client = {
       http.withCredentials = true
       http.onreadystatechange = function () {
         if (this.readyState === this.DONE) {
-          if (this.status === 200) {
+          if (this.status === 200 || this.status === 201) {
             return resolve(true)
           } else {
             reject({status: this.status, xhr: this})
@@ -228,9 +332,9 @@ var client = {
   }
 }
 
-// Alias some methods
-client.create = client.post
-client.replace = client.put
-client.update = client.patch
+// Alias some extra Solid web client methods
+SolidWebClient.create = SolidWebClient.post
+SolidWebClient.replace = SolidWebClient.put
+SolidWebClient.update = SolidWebClient.patch
 
-module.exports = client
+module.exports = SolidWebClient
